+++ temporal_sqlalchemy/__init__.py
@@ -7,5 +7,7 @@ from .bases import (
     TemporalProperty,
     TemporalActivityMixin)
 from .session import temporal_session, persist_history, is_temporal_session
-from .clock import add_clock, get_activity_clock_backref, get_history_model, get_history_model
+from .clock import add_clock
+from .util import get_activity_clock_backref, \
+    get_history_model
 from .core import TemporalModel
